Trying to refactor the code for fixing the cursor position out but rich text is too intertwined with readme. This is bad.
