+++ Editor/RichTextEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 061028dd89c64709baa7ce40d855aca8
+timeCreated: 1561127867
